refactor: PurgomalumClient 도메인 패키지로 이동 - PurgomalumClient 는 도메인 서비스이기에 도메인 패키지로 이동한다
