+++ src/main/java/kitchenpos/common/infra/DefaultPurgomalumClient.java
@@ -2,7 +2,7 @@ package kitchenpos.common.infra;
 
 import java.net.URI;
 
-import kitchenpos.common.application.PurgomalumClient;
+import kitchenpos.common.domain.PurgomalumClient;
 import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestTemplate;
