+++ src/test/java/kitchenpos/application/MenuServiceTest.java
@@ -7,7 +7,7 @@ import kitchenpos.menu.domain.MenuRepository;
 import kitchenpos.menu.application.MenuService;
 import kitchenpos.product.domain.Product;
 import kitchenpos.product.domain.ProductRepository;
-import kitchenpos.common.application.PurgomalumClient;
+import kitchenpos.common.domain.PurgomalumClient;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
