+++ src/test/java/kitchenpos/application/ProductServiceTest.java
@@ -4,7 +4,7 @@ import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuRepository;
 import kitchenpos.product.domain.Product;
 import kitchenpos.product.domain.ProductRepository;
-import kitchenpos.common.application.PurgomalumClient;
+import kitchenpos.common.domain.PurgomalumClient;
 import kitchenpos.product.application.ProductService;
 
 import org.junit.jupiter.api.BeforeEach;
